+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.1.0
+
+- Reduced Bundle size
+- Added - `font-awesome.settings.pngDimensions` settings. Set custom width and height to save icon as png.
+- Added - `font-awesome.settings.fillColor` settings. Set custom fill color when converting custom image to svg.
+
 ## v1.0.0
 
 - Initial Release